Moves navigation into own container
